requirements.txt
# AirScout - Python Dependencies
# ==============================

# Data Ingestion & Processing
pandas>=2.0.0,<3.0.0
sodapy>=2.2.0,<3.0.0

# Geospatial Processing
geopandas>=0.14.0,<1.0.0
shapely>=2.0.0,<3.0.0

# Database Connectivity
sqlalchemy>=2.0.0,<3.0.0
geoalchemy2>=0.14.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Streamlit Dashboard
streamlit>=1.28.0,<2.0.0
folium>=0.15.0,<1.0.0
streamlit-folium>=0.15.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0

# Push Notifications
pywebpush>=1.14.0,<2.0.0
cryptography>=41.0.0,<43.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
